requirements.txt
python-pkcs11
backoff
fastapi
uvicorn[standard]
pydantic
boto3